Use the same configuration then database for poll, rank and update services.
Configuration is shared through the config/ directory.  Having a shared
configuration make sure that all services are up to the same standard.  Of
course, if the frontend is rewritten in nodejs, it will not share the same
configuration than the other Python services.

A new docker compose file test.yml is responsible to test all servcies.  A simple:

	$ docker compose -f test.yml up --build --force-recreate

Will run all the test one by one and will stop as soon as one step failed.
Test can only be run for a particular service using the following command:

	$ docker compose -f test.yml up --build --force-recreate $service-test

Although it  may still run test for other services, it will only show the
result for the given service.  If the test for the other services fail
though, the test for the given service will not be run and the other services
will needs to be fixed before being able to run the test for the wanted
service.