Test for a bug in the Dgraph lexer.
There was a bug where the Dgraph lexer was not handling "\x00" for instance.
This commit introduce a tets to see if the bug is still happenning.

Turns out Dgraph not handles things just fine.  I think the bug came from the
gql python library, because after looking carefully to the GraphQL spec,
requests to GraphQL endpoints are JSON, and JSON deal with such escape
sequences in a way that don't broke Dgraph.